# docs/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/problems": {
      "get": {
        "description": "Returns a paginated list of problems. Staff users can pass all=1 to include hidden problems.",
        "operationId": "ProblemsController_list",
        "parameters": [
          {
            "name": "all",
            "required": false,
            "in": "query",
            "description": "Set to 1 to include hidden problems (staff only)",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "keyword",
            "required": false,
            "in": "query",
            "description": "Search keyword",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "tags",
            "required": false,
            "in": "query",
            "description": "Comma-separated tag filter",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "difficultyMin",
            "required": false,
            "in": "query",
            "description": "Minimum difficulty",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "difficultyMax",
            "required": false,
            "in": "query",
            "description": "Maximum difficulty",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of problems"
          }
        },
        "summary": "List problems",
        "tags": [
          "problems"
        ]
      },
      "post": {
        "operationId": "ProblemsController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateProblemDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Problem created"
          },
          "400": {
            "description": "Validation error"
          },
          "403": {
            "description": "Insufficient permissions"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create a new problem",
        "tags": [
          "problems"
        ]
      }
    },
    "/problems/import": {
      "post": {
        "description": "Upload a ZIP file containing problem definitions and testcases to bulk-import problems.",
        "operationId": "ProblemsController_importZip",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "file": {
                    "type": "string",
                    "format": "binary"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Problems imported successfully"
          },
          "400": {
            "description": "Invalid ZIP format"
          },
          "403": {
            "description": "Insufficient permissions"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Import problems from ZIP",
        "tags": [
          "problems"
        ]
      }
    },
    "/problems/{id}/export": {
      "get": {
        "operationId": "ProblemsController_exportZip",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Problem ID",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "testdata",
            "required": false,
            "in": "query",
            "description": "Set to \"false\" or \"0\" to exclude testdata",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "ZIP file download"
          },
          "403": {
            "description": "Insufficient permissions"
          },
          "404": {
            "description": "Problem not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Export problem as ZIP",
        "tags": [
          "problems"
        ]
      }
    },
    "/problems/{id}/testcases": {
      "get": {
        "operationId": "ProblemsController_listTestcases",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Problem ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of testcases"
          },
          "403": {
            "description": "Insufficient permissions"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List testcases for a problem",
        "tags": [
          "problems"
        ]
      },
      "post": {
        "operationId": "ProblemsController_createTestcase",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Problem ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTestcaseDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Testcase created"
          },
          "400": {
            "description": "Validation error"
          },
          "403": {
            "description": "Insufficient permissions"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create a testcase",
        "tags": [
          "problems"
        ]
      }
    },
    "/problems/{id}/testcases/{testcaseId}": {
      "get": {
        "operationId": "ProblemsController_getTestcase",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Problem ID",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "testcaseId",
            "required": true,
            "in": "path",
            "description": "Testcase ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Testcase detail including input/output"
          },
          "404": {
            "description": "Testcase not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get testcase detail",
        "tags": [
          "problems"
        ]
      },
      "patch": {
        "operationId": "ProblemsController_updateTestcase",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Problem ID",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "testcaseId",
            "required": true,
            "in": "path",
            "description": "Testcase ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateTestcaseDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Testcase updated"
          },
          "404": {
            "description": "Testcase not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update a testcase",
        "tags": [
          "problems"
        ]
      },
      "delete": {
        "operationId": "ProblemsController_deleteTestcase",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Problem ID",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "testcaseId",
            "required": true,
            "in": "path",
            "description": "Testcase ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Testcase deleted"
          },
          "404": {
            "description": "Testcase not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Delete a testcase",
        "tags": [
          "problems"
        ]
      }
    },
    "/problems/{id}": {
      "patch": {
        "operationId": "ProblemsController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Problem ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProblemDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Problem updated"
          },
          "404": {
            "description": "Problem not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update a problem",
        "tags": [
          "problems"
        ]
      }
    },
    "/problems/{number}/submit-context": {
      "get": {
        "description": "Returns problem metadata needed by the submission form, optionally scoped to a contest.",
        "operationId": "ProblemsController_submitContext",
        "parameters": [
          {
            "name": "number",
            "required": true,
            "in": "path",
            "description": "Problem number",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "contestId",
            "required": false,
            "in": "query",
            "description": "Contest ID for contest-scoped access",
            "schema": {}
          }
        ],
        "responses": {
          "200": {
            "description": "Submission context"
          },
          "404": {
            "description": "Problem not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get submission context for a problem",
        "tags": [
          "problems"
        ]
      }
    },
    "/problems/{number}/assets/{path}": {
      "get": {
        "operationId": "ProblemsController_getAsset",
        "parameters": [
          {
            "name": "number",
            "required": true,
            "in": "path",
            "description": "Problem number",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "path",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "contestId",
            "required": false,
            "in": "query",
            "description": "Contest ID for contest-scoped access",
            "schema": {}
          }
        ],
        "responses": {
          "200": {
            "description": "Asset file content"
          },
          "404": {
            "description": "Asset not found"
          }
        },
        "summary": "Get problem asset file (image, etc.)",
        "tags": [
          "problems"
        ]
      }
    },
    "/problems/{number}": {
      "get": {
        "description": "Returns full problem detail including statement, tags, and metadata. Contest-scoped if contestId is provided.",
        "operationId": "ProblemsController_detail",
        "parameters": [
          {
            "name": "number",
            "required": true,
            "in": "path",
            "description": "Problem number",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "contestId",
            "required": false,
            "in": "query",
            "description": "Contest ID for contest-scoped access",
            "schema": {}
          }
        ],
        "responses": {
          "200": {
            "description": "Problem detail"
          },
          "404": {
            "description": "Problem not found"
          }
        },
        "summary": "Get problem detail",
        "tags": [
          "problems"
        ]
      }
    },
...
    "/discussions/problems/{number}": {
      "post": {
        "operationId": "DiscussionsController_create",
        "parameters": [
          {
            "name": "number",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateDiscussionDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create a discussion post",
        "tags": [
          "discussions"
        ]
      },
      "get": {
        "operationId": "DiscussionsController_listByProblem",
        "parameters": [
          {
            "name": "number",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "default": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "default": 20,
              "type": "number"
            }
          },
          {
            "name": "sort",
            "required": false,
            "in": "query",
            "schema": {
              "default": "latest",
              "type": "string",
              "enum": [
                "latest",
                "popular"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "List discussions for a problem",
        "tags": [
          "discussions"
        ]
      }
    },
...
  },
  "info": {
    "title": "EpochJudge API",
    "description": "纪元 — 开源 Online Judge 平台 API",
    "version": "0.1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "CreateTestcaseDto": {
        "type": "object",
        "properties": {
          "input": {
            "type": "string"
          },
          "output": {
            "type": "string"
          },
          "score": {
            "type": "number"
          },
          "isSample": {
            "type": "boolean"
          },
          "ordinal": {
            "type": "number"
          }
        },
        "required": [
          "input",
          "output",
          "score"
        ]
      },
      "UpdateTestcaseDto": {
        "type": "object",
        "properties": {
          "input": {
            "type": "string"
          },
          "output": {
            "type": "string"
          },
          "score": {
            "type": "number"
          },
          "isSample": {
            "type": "boolean"
          },
          "ordinal": {
            "type": "number"
          }
        }
      },
      "CreateProblemDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string"
          },
          "statement": {
            "type": "string"
          },
          "difficulty": {
            "type": "number"
          },
          "visibility": {
            "type": "string",
            "enum": [
              "PUBLIC",
              "PRIVATE"
            ]
          },
          "timeLimitMs": {
            "type": "number"
          },
          "memoryLimitKb": {
            "type": "number"
          },
          "tags": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "title"
        ]
      },
      "UpdateProblemDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string"
          },
          "statement": {
            "type": "string"
          },
          "difficulty": {
            "type": "number"
          },
          "visibility": {
            "type": "string"
          },
          "timeLimitMs": {
            "type": "number"
          },
          "memoryLimitKb": {
            "type": "number"
          },
          "tags": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
...
      "CreateDiscussionDto": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "enum": [
              "QUESTION",
              "SOLUTION"
            ]
          },
          "title": {
            "type": "string"
          },
          "content": {
            "type": "string"
          }
        },
        "required": [
          "type",
          "title",
          "content"
        ]
      },
...
    }
  }
}